+++ src/shadertastic.cpp
@@ -65,6 +65,7 @@ bool module_loaded = false;
 //----------------------------------------------------------------------------------------------------------------------
 
 void load_effects(shadertastic_common *s, obs_data_t *settings, const std::string &effects_dir, const std::string &effects_type) {
+    UNUSED_PARAMETER(settings);
     std::vector<std::string> dirs = list_directories((effects_dir + "/" + effects_type + "s").c_str());
 
     for (const auto &dir : dirs) {